cmd: add argument for build environment command
Currently, `build-env` accepts both a spec and a command on the command
line. First, this is very hard to discover as the help output does not
mention it at all. Second, having to specify `--` to separate the spec
and command is not very convenient.

This change introduces a `-c`/`--command` argument that can be used to
specify the build environment command. The only downside is that we have
to split the command ourselves using `shlex`. The command argument is
now clearly documented in the help output, though.