Show status information in terminal title
Installing packages with a lot of dependencies does not have an easy way
of judging the current progress (apart from running `spack spec -I pkg`
in another terminal). This change makes Spack update the terminal's
title with status information, including the current installation phase
as well as information about the current and total number of packages.